Split prof from course table
